requirements.txt
mbrl==0.2.0
gymnasium==0.26.3
numpy<2.0
torch
matplotlib
interpret
scikit-learn
pandas
hydra-core
omegaconf
pyyaml
